Move Sort and Order into a separate file
