+++ core/src/main/scala/io/chrisdavenport/github/data/SearchResult.scala
@@ -14,59 +14,4 @@ object SearchResult {
       cursor.downField("items").as[List[A]]
     ).mapN(SearchResult.apply)
 
-  sealed trait Order
-
-  object Order {
-
-    case object Ascending extends Order
-    case object Descending extends Order
-
-    def toOptionalParam(order: Order): Option[String] =
-      order match {
-        case Ascending => Some("asc")
-        case Descending => Some("desc")
-      }
-
-  }
-
-
-  sealed trait Sort
-
-  object Sort {
-
-    case object BestMatch extends Repos.Sort with Users.Sort
-    case object Stars extends Repos.Sort
-    case object Forks extends Repos.Sort
-    case object HelpWantedIssues extends Repos.Sort
-    case object Updated extends Repos.Sort
-    case object Followers extends Users.Sort
-    case object Repositories extends Users.Sort
-    case object Joined extends Users.Sort
-
-    def toOptionalParam(sort: Sort): Option[String] =
-      sort match {
-        case BestMatch => None
-        case Stars => Some("stars")
-        case Forks => Some("forks")
-        case HelpWantedIssues => Some("help-wanted-issues")
-        case Updated => Some("updated")
-        case Followers => Some("followers")
-        case Repositories => Some("repositories")
-        case Joined => Some("joined")
-      }
-
-  }
-
-  object Repos {
-
-    sealed trait Sort extends SearchResult.Sort
-
-  }
-
-  object Users {
-
-    sealed trait Sort extends SearchResult.Sort
-
-  }
-
 }
